Fix the GuessSection bug
I believe this is a complete fix for the GuessSection bug.
The behaviour is now:
The path has to start with a "/" and a section is the first part of the
path. If the string "content" appears as a path element it is chopped.
If "content" appears it has to be the first non-empty string
A section can't have a "/" anywhere within it.
Which implies that the section is always the bit between the first
two "/"'s, after "content" has been removed.
We either return the section name or an empty string if it does
not meet the above

Currently all of the test cases pass, see path_test.go.